split colony sliders UI from logic
