+++ index.html
@@ -50,6 +50,7 @@
     <script src="colony.js"></script>
     <script src="colonyUI.js"></script>
     <script src="colonySliders.js"></script>
+    <script src="colonySlidersUI.js"></script>
     <script src="projects.js"></script>
     <script src="projectsUI.js"></script>
     <script src="spaceship.js"></script>
